# openapi.yaml
openapi: 3.0.3
info:
  title: pytheum API
  version: "0.1.0"
  description: |
    Point-in-time prediction-market context for AI agents. Four endpoints over
    a unified market + event index spanning Kalshi, Polymarket, and Manifold.
  contact:
    url: https://github.com/pytheum/pytheum-doc
servers:
  - url: https://api.pytheum.com
    description: Production
paths:
  /v1/markets/relevant-to:
    get:
      summary: Find markets matching a free-text query
      operationId: findRelevantMarkets
      parameters:
        - $ref: '#/components/parameters/Query'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/MinSimilarity'
        - $ref: '#/components/parameters/Venue'
        - $ref: '#/components/parameters/Status'
        - name: group_by
          in: query
          schema:
            type: string
            enum: [bundle, none]
            default: bundle
          description: |
            `bundle` (default): collapse to the highest-similarity market per
            bundle_id. `none`: return raw matches.
      responses:
        '200':
          description: Ranked list of markets matching the query
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MarketsResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '502':
          $ref: '#/components/responses/EmbeddingFailed'
...
components:
  parameters:
    Query:
      name: query
      in: query
      required: true
      schema:
        type: string
        maxLength: 4000
      description: Free-form text to embed and search against the market index
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    MinSimilarity:
      name: min_similarity
      in: query
      schema:
        type: number
        minimum: 0.0
        maximum: 1.0
      description: |
        Default varies by endpoint (0.40 for relevant-to, 0.15 for context,
        0.30 for events/related-markets). Cosine similarity floor.
    Venue:
      name: venue
      in: query
      schema:
        type: string
      description: Comma-separated list (kalshi,polymarket,manifold)
    Status:
      name: status
      in: query
      schema:
        type: string
        default: active,resolving
      description: Comma-separated list (active,resolving,resolved)
...
  schemas:
    Market:
      type: object
      properties:
        id: { type: string, example: "polymarket:31875" }
        question: { type: string, nullable: true }
        venue: { type: string, enum: [kalshi, polymarket, manifold] }
        bundle_id: { type: string, nullable: true }
        bundle_label: { type: string, nullable: true }
        similarity: { type: number, minimum: 0.0, maximum: 1.0 }
        volume_usd: { type: number, nullable: true }
        status: { type: string, enum: [active, resolving, resolved] }
        url: { type: string, nullable: true }
...
    Meta:
      type: object
      properties:
        embedding_model: { type: string, example: "openai:text-embedding-3-large" }
        limit: { type: integer }
        window_hours: { type: integer, example: 24, description: "context endpoints only" }
        total_in_window:
          type: integer
          description: |
            context endpoints only — number of events indexed in the rolling
            window. Currently capped at 8000 (FIFO drop when exceeded).
        min_similarity:
          type: number
          description: "context endpoints only — effective similarity floor used"
        group_by:
          type: string
          description: "market endpoints only — `bundle` or `none`"
        venues:
          type: array
          items: { type: string }
          nullable: true
          description: |
            market endpoints only — venues filter that was applied. `null`
            means no filter (default).
    MarketsResponse:
      type: object
      properties:
        query: { type: string, nullable: true }
        markets:
          type: array
          items: { $ref: '#/components/schemas/Market' }
        meta: { $ref: '#/components/schemas/Meta' }
...
    Error:
      type: object
      properties:
        detail: { type: string }
  responses:
    BadRequest:
      description: Missing or invalid parameter
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    EmbeddingFailed:
      description: Upstream embedding provider unavailable
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }